Added methods RegisterFolder and RegisterFile to MonitorableRepository.cs. Implemented them in LocalFileSystemRepository.cs, adding method PopulateFolder and a new overload of PopulateFile to Manifest.cs, and updated EnsureInitialStateOfLocalFiles in Program.cs to call them after making changes based on remote content that isn't found locally. Added centralized method SaveManifest to the base MonitorableRepository.cs class with implementations in GoogleDriveRepository.cs and LocalFileSystemRepository.cs. Replaced all direct calls to _manifest.SaveTo with calls to SaveManifest. Updated EnsureInitialStateOfLocalFiles() to call SaveManifest on the LocalFileSystemRepository instance. Fixed PopulateFile in Manifest.cs to set HasChanges to true. Updated the MoveFile method in LocalFileSystemRepository.cs to raise an exception if neither the old path nor the new path exist. If the new path exists and the old path doesn't, it is assumed the rename was already previously processed and it just didn't get recorded. If the old path exists, it is always moved to the new path, overwriting anything that is already there. Added class PathUtility to standardize path manipulation (especially the path separator character). Updated all manipulations of repository paths in GoogleDriveRepository.cs, LocalFileSystemRepository.cs and Manifest.cs to use the new methods. Reworked RemoveFile in GoogleDriveRepository.cs to move files to the trash instead of outright deleting them, so that if something does go unexpectedly wrong, it is possible to restore the files. Fixed MoveFile in GoogleDriveRepository.cs to create a blank slate File model instead of getting the existing state of the File in preparation for supplying an instance with a new name to Files.Update. Fixed RaiseChangedEvent in LocalFileSystemRepository.cs to supply repository paths instead of local filesystem paths for the OldFilePath field of the ChangeInfo instances for rename/move events. Fixed RequestChangeProcessorStop in Program.cs to pulse s_sync, waking the change processor thread if it is sleeping. Fixed the ChangeProcessorThreadProc loop to properly wake up and exit if s_changeProcesorStopping is true. Added additional shutdown logic and diagnostic output to Main in Program.cs.  Added ToString overrides to GoogleDriveRepository.cs and LocalFileSystemRepository.cs. Updated MusicBoxSynchronizer.csproj to set RollForward to Major, allowing the code to run on .NET 9.0 even though it targets net8.0.

+++ LocalFileSystemRepository.cs
@@ -8,6 +8,8 @@ namespace MusicBoxSynchronizer
 {
 	public class LocalFileSystemRepository : MonitorableRepository
 	{
+		public override string? ToString() => "Local Filesystem Interface";
+
 		public const string DefaultRepositoryPath = @"C:\MusicBoxDrive";
 
 		const string ManifestStateFileName = "local_drive_manifest";
@@ -75,12 +77,14 @@ namespace MusicBoxSynchronizer
 
 				var manifest = Manifest.Build(_rootPath!);
 
-				manifest.SaveTo(ManifestStateFileName);
+				manifest.HasChanges = true;
 
 				return manifest;
 			}
 
 			_manifest = TryLoadManifest() ?? BuildManifest();
+
+			SaveManifest();
 		}
 
 		public string RootPath => _rootPath;
@@ -177,9 +181,19 @@ namespace MusicBoxSynchronizer
 		{
 			OnDiagnosticOutput("Moving/renaming file: " + newPath + " (<- " + oldPath + ")");
 
+			string oldFullPath = GetFullPath(oldPath);
+			string newFullPath = GetFullPath(newPath);
+
+			bool oldExists = File.Exists(oldFullPath);
+			bool newExists = File.Exists(oldFullPath);
+
+			if (!oldExists && !newExists)
+				throw new Exception("Received file move/rename event but the file does not seem to exist: " + newPath + " (<- " + oldPath + ")");
+
 			File.Move(
 				GetFullPath(oldPath),
-				GetFullPath(newPath));
+				GetFullPath(newPath),
+				overwrite: true);
 		}
 
 		public override void RemoveFile(string path)
@@ -198,6 +212,26 @@ namespace MusicBoxSynchronizer
 			File.Copy(GetFullPath(path), temporaryLocation, overwrite: true);
 
 			return new TemporaryFileStream(temporaryLocation, FileMode.Open, FileAccess.Read);
+		}
+
+		public override void RegisterFolder(string path)
+		{
+			EnsureInitialized();
+
+			_manifest!.PopulateFolder(path, path);
+		}
+
+		public override void RegisterFile(ManifestFileInfo fileInfo)
+		{
+			EnsureInitialized();
+
+			_manifest!.PopulateFile(fileInfo, fileInfo.FilePath);
+		}
+
+		public override void SaveManifest()
+		{
+			if (_manifest?.HasChanges ?? false)
+				_manifest.SaveTo(ManifestStateFileName);
 		}
 
 		public override void StartMonitor()
@@ -269,7 +303,7 @@ namespace MusicBoxSynchronizer
 				fileSize,
 				modifiedTimeUTC);
 
-			_manifest.SaveTo(ManifestStateFileName);
+			SaveManifest();
 		}
 
 		void QueueChangedEvent(ChangeType changeType, string fullPath, string? oldFullPath = null)
@@ -395,8 +429,8 @@ namespace MusicBoxSynchronizer
 					return;
 			}
 
-			string relativePath = fullPath.Substring(_rootPath.Length + 1);
-			string? oldRelativePath = (oldFullPath == null) ? null : oldFullPath.Substring(_rootPath.Length + 1);
+			string relativePath = PathUtility.GetRelativePath(_rootPath, fullPath) ?? throw new Exception("Unable to determine relative path for full path: " + fullPath);
+			string? oldRelativePath = PathUtility.GetRelativePath(_rootPath, oldFullPath);
 
 			string md5Checksum = File.Exists(fullPath) ? MD5Utility.ComputeChecksum(fullPath) : "-";
 
@@ -407,7 +441,7 @@ namespace MusicBoxSynchronizer
 					changeType: ChangeType.Removed,
 					filePath: relativePath));
 			}
-			else if (oldFullPath == null)
+			else if (oldRelativePath == null)
 			{
 				OnChangeDetected(new ChangeInfo(
 					sourceRepository: this,
@@ -422,7 +456,7 @@ namespace MusicBoxSynchronizer
 					sourceRepository: this,
 					changeType: changeType,
 					filePath: relativePath,
-					oldFilePath: oldFullPath,
+					oldFilePath: oldRelativePath,
 					isFolder: Directory.Exists(fullPath)));
 			}
 		}
